Moving features in features folder
- All features are moved to features/
- The sub-packages loose their underscore

Everything is still callable through the old fst/ path
